+++ app/src/main/java/com/hippo/ehviewer/ui/DirPickerActivity.java
@@ -151,15 +151,4 @@ public class DirPickerActivity extends ToolbarActivity
             mPath.setText(dir.getPath());
         }
     }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        if(Settings.getEnabledSecurity()){
-            getWindow().setFlags(WindowManager.LayoutParams.FLAG_SECURE,
-                    WindowManager.LayoutParams.FLAG_SECURE);
-        }else{
-            getWindow().clearFlags(WindowManager.LayoutParams.FLAG_SECURE);
-        }
-    }
 }
