+++ app/src/main/java/com/hippo/ehviewer/ui/GalleryActivity.java
@@ -407,12 +407,6 @@ public class GalleryActivity extends EhActivity implements SeekBar.OnSeekBarChan
         if (mGLRootView != null) {
             mGLRootView.onResume();
         }
-        if(Settings.getEnabledSecurity()){
-            getWindow().setFlags(WindowManager.LayoutParams.FLAG_SECURE,
-                    WindowManager.LayoutParams.FLAG_SECURE);
-        }else{
-            getWindow().clearFlags(WindowManager.LayoutParams.FLAG_SECURE);
-        }
     }
 
     @Override
